requirements/mac_req.txt
absl-py==1.4.0
alembic==1.10.3
appnope==0.1.3
array-record==0.2.0
asttokens==2.2.1
astunparse==1.6.3
Authlib==1.2.0
backcall==0.2.0
bottle==0.12.25
brotlipy==0.7.0
cachetools==5.3.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
colorlog==6.7.0
conda==23.3.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.8
etils==1.3.0
executing==1.2.0
filelock==3.11.0
flatbuffers==1.12
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
grpcio==1.53.0
imageio==2.27.0
importlib-resources==5.12.0
ipython==8.13.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.0.4
mpmath==1.3.0
networkx==3.1
numpy==1.23.0
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optuna==3.1.1
optuna-dashboard==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.15.0
SQLAlchemy==2.0.9
stack-data==0.6.2
style==1.1.0
sympy==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.20.0
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.0
tensorflow-metadata==1.13.1
tensorflow-metal==0.5.0
tensorflow-probability==0.13.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.21
toml==0.10.2
torch==2.0.0
traitlets==5.9.0
tqdm
typeguard==2.13.3
typing_extensions==4.5.0
update==0.0.1
validators==0.20.0
wcwidth==0.2.6
weaviate-client==3.19.2
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
